Minor tweaks and refactors
- New setup folder
- Fixes for some comments in pr
