+++ bot/setup/command.py
